rename test, quasi_incompressible + fix single #
